requirements.txt
# For local development
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
Jinja2==3.1.6
mangum==0.19.0
tensorflow==2.19.0
numpy==2.0.0
scipy==1.15.2
pillow==11.1.0
